Update To do list

+++ Quarto-periodo/Gestao-de-Ciclo-de-Vida-da-Aplicacao/TDD/src/main/java/org/example/Calculator.java
@@ -12,7 +12,7 @@ public class Calculator {
     public int mult (int a, int b) {
         return a * b;
     }
-    public int divide(int a, int b) {
-        return a / b;
+    public float divide(int a, int b) {
+        return (float) a / b;
     }
 }
